+++ www/go/core/install/install.sql
@@ -112,7 +112,8 @@ CREATE TABLE `core_customfields_field` (
   `prefix` varchar(32) NOT NULL DEFAULT '',
   `suffix` varchar(32) NOT NULL DEFAULT '',
   `options` text DEFAULT NULL,
-  `hiddenInGrid` BOOLEAN NOT NULL DEFAULT TRUE
+  `hiddenInGrid` BOOLEAN NOT NULL DEFAULT TRUE,
+  `filterable` BOOLEAN NOT NULL DEFAULT FALSE
 ) ENGINE=InnoDB;
 
 CREATE TABLE `core_customfields_field_set` (
@@ -776,7 +777,8 @@ CREATE TABLE `core_entity_filter` (
   `name` varchar(190) COLLATE utf8mb4_unicode_ci NOT NULL,
   `createdBy` int(11) NOT NULL,
   `filter` text COLLATE utf8mb4_unicode_ci NOT NULL,
-  `aclId` int(11) NOT NULL
+  `aclId` int(11) NOT NULL,
+  `type` ENUM('fixed','variable') NOT NULL DEFAULT 'fixed'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci ROW_FORMAT=COMPACT;