+++ www/go/core/model/EntityFilter.php
@@ -57,11 +57,14 @@ class EntityFilter extends AclOwnerEntity {
 
 	protected static function defineFilters() {
 		return parent::defineFilters()
-						->add('entity', function (Criteria $criteria, $value, \go\core\orm\Query $query){
-							$query->join('core_entity', 'e', 'e.id = f.entityTypeId');
-							
-							$criteria->where(['e.clientName' => $value]);
-						});
+			->add('entity', function (Criteria $criteria, $value, \go\core\orm\Query $query){
+				$query->join('core_entity', 'e', 'e.id = f.entityTypeId');
+
+				$criteria->where(['e.clientName' => $value]);
+			})
+			->add('type', function (Criteria $criteria, $value) {
+				$criteria->where('type','=', $value);
+			});
 	}
 
 	public static function sort(\go\core\orm\Query $query, array $sort)